fix: prevent electron.app from being undefined in tests
Before, we had an import chain from the test page object `MenuBar.ts` -> `fileMenu.ts` -> `listeners.ts` -> `opossum-file.ts`, which creates a LegacyFileConverter on import, which calls `FileConverter.getOpossumFileExecutable`, which depends on `electron.app`. But because that is in a test context, `electron.app` is not defined.

Now, I moved `importFileFormats` to `shared-constants`, so `MenuBar` doesn't need to import from electron code. So we can now depend on `app` not being `undefined` in `FileConverter.ts
`.
