+++ src/ElectronBackend/main/__tests__/menu.test.ts
@@ -15,6 +15,7 @@ vi.mock('electron', () => {
   const mockElectron = {
     BrowserWindow: class BrowserWindowMock {},
     app: {
+      getAppPath: () => './',
       isPackaged: true,
     },
     Menu: {
